requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
ddgs==9.10.0
httpx==0.28.1
pillow==12.1.0
